"AddMultiplteWorkOverRecompletion"

+++ Backend_UMR_Work_Program/Controllers/WorkProgrammeController.cs
@@ -4323,7 +4323,7 @@ namespace Backend_UMR_Work_Program.Controllers
 
 
         [HttpPost("POST_WORKOVERS_RECOMPLETION_JOB")]
-        public async Task<object> POST_WORKOVERS_RECOMPLETION_JOB([FromBody] WORKOVERS_RECOMPLETION_JOB1 workovers_recompletion_model, string omlName, string fieldName, string year, string actionToDo)
+        public async Task<object> POST_WORKOVERS_RECOMPLETION_JOB([FromBody] WORKOVERS_RECOMPLETION_JOB1 workovers_recompletion_model, string omlName, string fieldName, string year, string actionToDo, int? id)
         {
 
             int save = 0;
@@ -4331,56 +4331,71 @@ namespace Backend_UMR_Work_Program.Controllers
 
             WORKOVERS_RECOMPLETION_JOB1 getData;
 
+           
+
             try
             {
-                #region Saving FDP data
-                if (workovers_recompletion_model != null)
+                if (id > 0 || workovers_recompletion_model != null)
                 {
-                    if (concessionField?.Field_Name != null)
-                    {
-                        getData = await (from c in _context.WORKOVERS_RECOMPLETION_JOBs1 where c.OML_Name == omlName && c.Field_ID == concessionField.Field_ID && c.COMPANY_ID == WKPCompanyId && c.Year_of_WP == year && c.QUATER == workovers_recompletion_model.QUATER select c).FirstOrDefaultAsync();
 
-                    }
-                    else
+                    if (id > 0 && workovers_recompletion_model.Id==0)
                     {
-                        getData = await (from c in _context.WORKOVERS_RECOMPLETION_JOBs1 where c.OML_Name == omlName && c.COMPANY_ID == WKPCompanyId && c.Year_of_WP == year && c.QUATER == workovers_recompletion_model.QUATER select c).FirstOrDefaultAsync();
+                        getData = await (from c in _context.WORKOVERS_RECOMPLETION_JOBs1 where c.Id == id select c).FirstOrDefaultAsync();
 
+                        _context.WORKOVERS_RECOMPLETION_JOBs1.Remove(getData);
+                        action = GeneralModel.Delete;
                     }
 
-                    workovers_recompletion_model.Companyemail = WKPCompanyEmail;
-                    workovers_recompletion_model.CompanyName = WKPCompanyName;
-                    workovers_recompletion_model.COMPANY_ID = WKPCompanyId;
-                    workovers_recompletion_model.CompanyNumber = WKPCompanyNumber;
-                    workovers_recompletion_model.Year_of_WP = year;
-                    workovers_recompletion_model.OML_Name = omlName.ToUpper();
-                    workovers_recompletion_model.Field_ID = concessionField?.Field_ID ?? null;
-                    //workovers_recompletion_model.Actual_year = year;
-                    //	workovers_recompletion_model.proposed_year = (int.Parse(year) + 1).ToString();
-                    workovers_recompletion_model.proposed_year = year;
-
-                    if (action == GeneralModel.Insert)
+                    #region Saving FDP data
+                    else if (workovers_recompletion_model != null)
                     {
-                        if (getData == null)
+                        if (concessionField?.Field_Name != null)
                         {
-                            workovers_recompletion_model.Date_Created = DateTime.Now;
-                            workovers_recompletion_model.Created_by = WKPCompanyId;
-                            await _context.WORKOVERS_RECOMPLETION_JOBs1.AddAsync(workovers_recompletion_model);
+                            getData = await (from c in _context.WORKOVERS_RECOMPLETION_JOBs1 where c.OML_Name == omlName && c.Field_ID == concessionField.Field_ID && c.COMPANY_ID == WKPCompanyId && c.CompletionWellName.ToUpper().Trim() == workovers_recompletion_model.CompletionWellName.ToUpper().Trim() && c.Year_of_WP == year && c.QUATER == workovers_recompletion_model.QUATER select c).FirstOrDefaultAsync();
+
                         }
                         else
                         {
-                            workovers_recompletion_model.Date_Created = getData.Date_Created;
-                            workovers_recompletion_model.Created_by = getData.Created_by;
-                            workovers_recompletion_model.Date_Updated = DateTime.Now;
-                            workovers_recompletion_model.Updated_by = WKPCompanyId;
-                            _context.WORKOVERS_RECOMPLETION_JOBs1.Remove(getData);
-                            await _context.WORKOVERS_RECOMPLETION_JOBs1.AddAsync(workovers_recompletion_model);
-                        }
-                    }
-                    else if (action == GeneralModel.Delete)
-                    {
-                        _context.WORKOVERS_RECOMPLETION_JOBs1.Remove(getData);
-                    }
+                            getData = await (from c in _context.WORKOVERS_RECOMPLETION_JOBs1 where c.OML_Name == omlName && c.COMPANY_ID == WKPCompanyId && c.CompletionWellName.ToUpper().Trim() == workovers_recompletion_model.CompletionWellName.ToUpper().Trim() && c.Year_of_WP == year && c.QUATER == workovers_recompletion_model.QUATER select c).FirstOrDefaultAsync();
 
+                        }
+
+                        workovers_recompletion_model.Companyemail = WKPCompanyEmail;
+                        workovers_recompletion_model.CompanyName = WKPCompanyName;
+                        workovers_recompletion_model.COMPANY_ID = WKPCompanyId;
+                        workovers_recompletion_model.CompanyNumber = WKPCompanyNumber;
+                        workovers_recompletion_model.Year_of_WP = year;
+                        workovers_recompletion_model.OML_Name = omlName.ToUpper();
+                        workovers_recompletion_model.Field_ID = concessionField?.Field_ID ?? null;
+                        //workovers_recompletion_model.Actual_year = year;
+                        //	workovers_recompletion_model.proposed_year = (int.Parse(year) + 1).ToString();
+                        workovers_recompletion_model.proposed_year = year;
+
+                        if (action == GeneralModel.Insert)
+                        {
+                            if (getData == null)
+                            {
+                                workovers_recompletion_model.Date_Created = DateTime.Now;
+                                workovers_recompletion_model.Created_by = WKPCompanyId;
+                                await _context.WORKOVERS_RECOMPLETION_JOBs1.AddAsync(workovers_recompletion_model);
+                            }
+                            else
+                            {
+                                workovers_recompletion_model.Date_Created = getData.Date_Created;
+                                workovers_recompletion_model.Created_by = getData.Created_by;
+                                workovers_recompletion_model.Date_Updated = DateTime.Now;
+                                workovers_recompletion_model.Updated_by = WKPCompanyId;
+                                _context.WORKOVERS_RECOMPLETION_JOBs1.Remove(getData);
+                                await _context.WORKOVERS_RECOMPLETION_JOBs1.AddAsync(workovers_recompletion_model);
+                            }
+                        }
+                        else if (action == GeneralModel.Delete)
+                        {
+                            _context.WORKOVERS_RECOMPLETION_JOBs1.Remove(getData);
+                        }
+
+
+                    }
                     save += await _context.SaveChangesAsync();
 
                     if (save > 0)
@@ -4395,8 +4410,9 @@ namespace Backend_UMR_Work_Program.Controllers
                         return BadRequest(new { message = "Error : An error occured while trying to submit this form." });
 
                     }
-                }
 
+
+                }
                 return BadRequest(new { message = $"Error : No data was passed for {actionToDo} process to be completed." });
                 #endregion
